+++ src/routes/index.ts
@@ -1,0 +1,8 @@
+import express, { Request, Response } from 'express';
+const router = express.Router();
+
+router.get('/', (req: Request, res: Response) => {
+    res.render('index');
+});
+
+export default router;
