+++ tools/copyAssets.ts
@@ -1,0 +1,5 @@
+import * as shell from "shelljs";
+
+// Copy all the view templates
+shell.cp("-R", "src/views", "build/");
+shell.cp("-R", "src/public", "build/");
